move ClauseList out of Criterion
